+++ packages/sdk/src/exceptions/handler.js
@@ -36,7 +36,6 @@ function handler(error) {
             const {
                 code,
                 message,
-                status,
                 statusText,
             } = (details || {});
